Test more edges of invalid input

+++ micro_namedtuple_sqlite_persister/persister.py
@@ -50,7 +50,7 @@ class Engine:
 
     def update(self, row: NamedTuple) -> None:
         if row[0] is None:
-            raise ValueError("Cannot update row when id=None")
+            raise ValueError("Cannot UPDATE, id=None")
         query = f"""
             UPDATE {row.__class__.__name__}
             SET {', '.join(f"{f} = ?" for f in row._fields)}
@@ -58,25 +58,34 @@ class Engine:
             """
         cur = self.connection.execute(query, (*row, row[0]))
         if cur.rowcount == 0:
-            raise ValueError(f"No row with id={row[0]} in table `{row.__class__.__name__}`")
+            raise ValueError(f"Cannot UPDATE, no row with id={row[0]} in table `{row.__class__.__name__}`")
         self.connection.commit()
 
-    def delete(self, Model: type[ROW], row_id: int) -> None:
+    def delete(self, Model: type[ROW], row_id: int | None) -> None:
+        if row_id is None:
+            raise ValueError("Cannot DELETE, id=None")
         query = f"""
             DELETE FROM {Model.__name__}
             WHERE id = ?
             """
-        self.connection.execute(query, (row_id,))
+        cur = self.connection.execute(query, (row_id,))
+        if cur.rowcount == 0:
+            raise ValueError(f"Cannot DELETE, no row with id={row_id} in table `{Model.__name__}`")
 
     ##### Reading
-    def get(self, Model: type[ROW], row_id: int) -> ROW:
+    def get(self, Model: type[ROW], row_id: int | None) -> ROW:
+        if row_id is None:
+            raise ValueError("Cannot SELECT, id=None")
         query = f"""
             SELECT {', '.join(Model._fields)}
             FROM {Model.__name__}
             WHERE id = ?
             """
         cursor = self.connection.execute(query, (row_id,))
-        return Model._make(cursor.fetchone())
+        row = cursor.fetchone()
+        if row is None:
+            raise ValueError(f"Cannot SELECT, no row with id={row_id} in table `{Model.__name__}`")
+        return Model._make(row)
 
 
 ## Serialization and Deserialization
